# Disable ImageBitmap in Safari