--- a/src/framework/parsers/texture/img-alpha-test.js
+++ b/src/framework/parsers/texture/img-alpha-test.js
@@ -83,6 +83,4 @@ class ImgAlphaTest {
     }
 }
 
-export {
-    ImgAlphaTest
-};
+export { ImgAlphaTest };