--- a/src/framework/parsers/texture/img.js
+++ b/src/framework/parsers/texture/img.js
@@ -5,6 +5,9 @@ import { Texture } from '../../../platform/graphics/texture.js';
 import { http } from '../../../platform/net/http.js';
 
 import { ABSOLUTE_URL } from '../../asset/constants.js';
+import { ImgAlphaTest } from './img-alpha-test.js';
+import { Debug } from '../../../core/debug.js';
+import { Tracing } from '../../../core/tracing.js';
 
 /** @typedef {import('../../handlers/texture.js').TextureParser} TextureParser */
 
@@ -20,6 +23,13 @@ class ImgParser {
         this.crossOrigin = registry.prefix ? 'anonymous' : null;
         this.maxRetries = 0;
         this.device = device;
+
+        // run image alpha test
+        Debug.call(() => {
+            if (Tracing.get('IMG_ALPHA_TEST')) {
+                ImgAlphaTest.run(this.device);
+            }
+        });
     }
 
     load(url, callback, asset) {