--- a/src/platform/graphics/webgl/webgl-graphics-device.js
+++ b/src/platform/graphics/webgl/webgl-graphics-device.js
@@ -1,5 +1,6 @@
 import { setupVertexArrayObject } from '../../../polyfill/OESVertexArrayObject.js';
 import { Debug } from '../../../core/debug.js';
+import { Tracing } from '../../../core/tracing.js';
 import { platform } from '../../../core/platform.js';
 import { Color } from '../../../core/math/color.js';
 
@@ -29,6 +30,7 @@ import { Texture } from '../texture.js';
 import { DebugGraphics } from '../debug-graphics.js';
 
 import { WebglVertexBuffer } from './webgl-vertex-buffer.js';
+import { WebglImageTest } from './wegbl-image-test.js';
 import { WebglIndexBuffer } from './webgl-index-buffer.js';
 import { WebglShader } from './webgl-shader.js';
 import { WebglTexture } from './webgl-texture.js';
@@ -236,52 +238,6 @@ function testTextureFloatHighPrecision(device) {
     return f === 0;
 }
 
-// ImageBitmap current state (Sep 2022):
-// - Lastest Chrome and Firefox browsers appear to support the ImageBitmap API fine (though
-//   there are likely still issues with older versions of both).
-// - Safari supports the API, but completely destroys some pngs. For example the cubemaps in
-//   steampunk slots https://playcanvas.com/editor/scene/524858. See the webkit issue
-//   https://bugs.webkit.org/show_bug.cgi?id=182424 for status.
-// - Some applications assume that PNGs loaded by the engine use HTMLImageBitmap interface and
-//   fail when using ImageBitmap. For example, Space Base project fails because it uses engine
-//   texture assets on the dom https://playcanvas.com/editor/scene/446278.
-
-// This function tests whether the current browser destroys PNG data or not.
-function testImageBitmap(device) {
-    // 1x1 png image containing rgba(1, 2, 3, 63)
-    const pngBytes = new Uint8Array([
-        137, 80, 78, 71, 13, 10, 26, 10, 0, 0, 0, 13, 73, 72, 68, 82, 0, 0, 0, 1, 0, 0, 0, 1, 8, 6, 0, 0, 0, 31, 21,
-        196, 137, 0, 0, 0, 13, 73, 68, 65, 84, 120, 218, 99, 100, 100, 98, 182, 7, 0, 0, 89, 0, 71, 67, 133, 148, 237,
-        0, 0, 0, 0, 73, 69, 78, 68, 174, 66, 96, 130
-    ]);
-
-    return createImageBitmap(new Blob([pngBytes], { type: 'image/png' }), { premultiplyAlpha: 'none' })
-        .then((image) => {
-            // create the texture
-            const texture = new Texture(device, {
-                width: 1,
-                height: 1,
-                format: PIXELFORMAT_RGBA8,
-                mipmaps: false,
-                levels: [image]
-            });
-
-            // read pixels
-            const rt = new RenderTarget({ colorBuffer: texture, depth: false });
-            device.setFramebuffer(rt.impl._glFrameBuffer);
-            device.initRenderTarget(rt);
-
-            const data = new Uint8ClampedArray(4);
-            device.gl.readPixels(0, 0, 1, 1, device.gl.RGBA, device.gl.UNSIGNED_BYTE, data);
-
-            rt.destroy();
-            texture.destroy();
-
-            return data[0] === 1 && data[1] === 2 && data[2] === 3 && data[3] === 63;
-        })
-        .catch(e => false);
-}
-
 /**
  * The graphics device manages the underlying graphics context. It is responsible for submitting
  * render state changes and graphics primitives to the hardware. A graphics device is tied to a
@@ -386,25 +342,29 @@ class WebglGraphicsDevice extends GraphicsDevice {
             Debug.log("Antialiasing has been turned off due to rendering issues on AppleWebKit 15.4");
         }
 
-        // Retrieve the WebGL context
-        const preferWebGl2 = (options.preferWebGl2 !== undefined) ? options.preferWebGl2 : true;
-
-        const names = preferWebGl2 ? ["webgl2", "webgl", "experimental-webgl"] : ["webgl", "experimental-webgl"];
         let gl = null;
-        for (let i = 0; i < names.length; i++) {
-            gl = canvas.getContext(names[i], options);
 
-            if (gl) {
-                this.webgl2 = (names[i] === DEVICETYPE_WEBGL2);
-                this._deviceType = this.webgl2 ? DEVICETYPE_WEBGL2 : DEVICETYPE_WEBGL1;
-                break;
+        // Retrieve the WebGL context
+        if (options.gl) {
+            gl = options.gl;
+        } else {
+            const preferWebGl2 = (options.preferWebGl2 !== undefined) ? options.preferWebGl2 : true;
+            const names = preferWebGl2 ? ["webgl2", "webgl", "experimental-webgl"] : ["webgl", "experimental-webgl"];
+            for (let i = 0; i < names.length; i++) {
+                gl = canvas.getContext(names[i], options);
+                if (gl) {
+                    break;
+                }
             }
         }
-        this.gl = gl;
 
         if (!gl) {
             throw new Error("WebGL not supported");
         }
+
+        this.gl = gl;
+        this.webgl2 = gl instanceof WebGL2RenderingContext;
+        this._deviceType = this.webgl2 ? DEVICETYPE_WEBGL2 : DEVICETYPE_WEBGL1;
 
         // pixel format of the framebuffer
         const alphaBits = gl.getParameter(gl.ALPHA_BITS);
@@ -412,12 +372,13 @@ class WebglGraphicsDevice extends GraphicsDevice {
 
         const isChrome = platform.browser && !!window.chrome;
         const isMac = platform.browser && navigator.appVersion.indexOf("Mac") !== -1;
+        const isSafari = platform.browser && !!window.safari;
 
         // enable temporary texture unit workaround on desktop safari
-        this._tempEnableSafariTextureUnitWorkaround = platform.browser && !!window.safari;
+        this._tempEnableSafariTextureUnitWorkaround = isSafari;
 
         // enable temporary workaround for glBlitFramebuffer failing on Mac Chrome (#2504)
-        this._tempMacChromeBlitFramebufferWorkaround = isMac && isChrome && !options.alpha;
+        this._tempMacChromeBlitFramebufferWorkaround = isMac && isChrome && !gl.alpha;
 
         // init polyfill for VAOs under webgl1
         if (!this.webgl2) {
@@ -432,13 +393,15 @@ class WebglGraphicsDevice extends GraphicsDevice {
         this.initializeRenderState();
         this.initializeContextCaches();
 
-        // start async image bitmap test
-        this.supportsImageBitmap = null;
-        if (typeof ImageBitmap !== 'undefined') {
-            testImageBitmap(this).then((result) => {
-                this.supportsImageBitmap = result;
-            });
-        }
+        // only enable ImageBitmap on chrome
+        this.supportsImageBitmap = isChrome;
+
+        // run image test
+        Debug.call(() => {
+            if (Tracing.get('RUN_IMAGE_ALPHA_TEST')) {
+                WebglImageTest.run(this);
+            }
+        });
 
         this.glAddress = [
             gl.REPEAT,