ImageBitmap only loads images faster than HTMLImageElement on Chrome.

Instead of performing an async canvas test for ImageBitmap at startup, we instead simply limit it to Chrome. The async startup test was returning too slowly in some applications and so wasn't reliable anyway.

For future testing, the alpha test functionality has been kept, though moved out of the graphics device. The test can be performed at startup in debug build by enabling the trace `IMG_ALPHA_TEST`.