@slimbuck 2023-04-11
So we don't currently use ImageBitmap on Safari at all, because the startup alpha test returns negative. However on Chrome, the startup test for ImageBitmap sometimes returns after textures have already started loading, meaning that in some cases Chrome -isn't- using ImageBitmap when it should.

Regarding your points above:
- If startup time is a priority, then faster loading times wins over main thread locking. I haven't tested thread locking performance on Safari because the profiler doesn't show this level of info, but it also doesn't matter because Safari premultiplies alpha anyway (rendering ImageBitmap mostly useless).
- We can update the test, but it currently just checks for Chrome.
- You're right, but I don't see a way to fix this. In an ideal world we'd have a test that tells us whether ImageBitmap works correctly and how it performs compared to HTMLImageElement. The test wouldn't take any time.
- Here are average timings on MacOS loading a single gltf with 6 textures in jpg & png format (first time is using HTMLImageBitmap, second is using ImageBitmap):
Safari - 550ms -> 650ms.
Firefox - 230ms -> 270ms.
Chrome - 310ms -> 280ms.
These timings include creating the device rendering the first frame.
- I thought there already was, but don't see an issue for the RGB premul issue. I can create a simple repro and report it.